requirements.txt
pywin32>=306
customtkinter>=5.2.0
matplotlib>=3.8.0
psutil>=5.9.0
pyinstaller>=6.0.0
pystray>=0.19.5
Pillow>=10.0.0

